[android-11.0.0_r20] DO NOT MERGE Attribute SOURCE_HOTWORD to OP_RECORD_AUDIO_HOTWORD.
Whenever we startRecording / finishRecording with a source
of SOURCE_HOTWORD, attribute them to OP_RECORD_AUDIO_HOTWORD
instead of OP_RECORD_AUDIO. In either case, OP_RECORD_AUDIO
is checked before the recording commences. Note that we also
check that all recordings with SOURCE_HOTWORD are initiated
by a uid that holds the privileged CAPTURE_AUDIO_HOTWORD
permission.

Note that this change will be superseded in the future once
we have stronger guarantees around hotword handling.

Bug: 162547999
Test: manual with dumpsys appops on AGSA

Change-Id: I70e0b652c357597770a18a5a385d62668a69059a